Add a blacklist to the scheduler
The blacklist file (default /etc/icecc/blacklist.conf) lists
names of clients that the scheduler will never consider for
running compiler jobs.

It is possible to update the blacklist file while the scheduler
is running, which makes it a convenient tool for manually blocking
misbehaving clients in a centrlaized manner, without having to
restart the scheduler.